Separated out packaging instructions and standards into child classes of a generic PackagingCodes classs. Moved database handling of packaging standards to the Standards class
