+++ cfr_tool/packaging_codes.py
@@ -1,6 +1,5 @@
 import networkx as nx
 import regex as re
-
 import clean_text as ct
 
 
@@ -11,63 +10,13 @@ Change this to be a class tha represents packaging standards in general.
 Convert specific subparts to be children of this class.
 '''
 
-class PackagingStandards:
-    PART = 178 
+class PackagingCodes:
 
     def __init__(self, db, soup):
         self.db = db
         self.soup = soup
         self.categories = []
     
-
-    def create_kinds_table(self):
-        self.db.execute('''
-            CREATE TABLE IF NOT EXISTS packaging_kinds(
-                id_code integer,
-                meaning text
-            );
-        ''')
-
-    def create_materials_table(self):
-        self.db.execute('''
-            CREATE TABLE IF NOT EXISTS packaging_materials (
-                id_code integer,
-                meaning text
-            );
-        ''') 
-
-    def create_categories_table(self):
-        self.db.execute('''
-            DROP TABLE IF EXISTS packaging_categories;
-        ''')
-        self.db.execute('''
-            CREATE TABLE packaging_categories (
-                full_code varchar not null primary key,
-                kind_id integer,
-                material_id text,
-                category_id integer,
-                category_desc text,
-                FOREIGN KEY (kind_id)
-                    REFERENCES packaging_kinds (id_code),
-                FOREIGN KEY (material_id)
-                    REFERENCES packaging_materials (id_code)
-            );
-        ''')
-
-    def load_packaging_categories(self):
-        self.create_categories_table()
-        self.db.executemany('''
-            INSERT INTO packaging_categories (
-                full_code,
-                kind_id,
-                material_id,
-                category_id,
-                category_desc
-            ) VALUES (
-                ?, ?, ?, ?, ?
-            )
-        ''', self.categories)
-
     def get_codes_descriptions(self, start, end, definition_paragraph='a'):
         '''
         Extracts packaging codes and the associated text in its tag
