+++ cfr_tool/standards.py
@@ -1,0 +1,59 @@
+from .packaging_codes import PackagingCodes
+from .soup import Soup
+
+class Standards(PackagingCodes):
+    PART = 178
+    START = 500
+    END = 523
+    def __init__(self, db, soup):
+        PackagingCodes.__init__(self, db, soup)
+        self.code_descriptions = self.get_codes_descriptions(self.START, self.END)
+
+    
+    def create_kinds_table(self):
+        self.db.execute('''
+            CREATE TABLE IF NOT EXISTS packaging_kinds(
+                id_code integer,
+                meaning text
+            );
+        ''')
+
+    def create_materials_table(self):
+        self.db.execute('''
+            CREATE TABLE IF NOT EXISTS packaging_materials (
+                id_code integer,
+                meaning text
+            );
+        ''') 
+
+    def create_categories_table(self):
+        self.db.execute('''
+            DROP TABLE IF EXISTS packaging_categories;
+        ''')
+        self.db.execute('''
+            CREATE TABLE packaging_categories (
+                full_code varchar not null primary key,
+                kind_id integer,
+                material_id text,
+                category_id integer,
+                category_desc text,
+                FOREIGN KEY (kind_id)
+                    REFERENCES packaging_kinds (id_code),
+                FOREIGN KEY (material_id)
+                    REFERENCES packaging_materials (id_code)
+            );
+        ''')
+
+    def load_packaging_categories(self):
+        self.create_categories_table()
+        self.db.executemany('''
+            INSERT INTO packaging_categories (
+                full_code,
+                kind_id,
+                material_id,
+                category_id,
+                category_desc
+            ) VALUES (
+                ?, ?, ?, ?, ?
+            )
+        ''', self.categories)
